Add files via upload
I am creating a “Wormhole” game for kids to evoke the feeling of being more interested in both the universe and physic and being able to understand the basic physic of wormhole – distort space and make a shortcut from one part of the universe to another part. In this game, kids can arrange the position of the wormhole as they want as long as the final bomb can go through the wormhole and hit the space monster.